requirements/requirements.txt
GitPython==2.1.10
docker==3.3.0